Create Stuff/ directory for all elm-stuff/ functionality
Combine Deps.Validate and Deps.Info into Stuff.Validate. The resulting
DepsInfo will be an “access token” to use any other part of elm-stuff/
to make sure it is not called before you validate things.

Also, make elm-stuff/ the name for the cache directory. Remove
configurability here. It’s not worth it.