Create Stuff.Interfaces for safely loading dependency interfaces
Under normal circumstances, this will not change. So you can read it
all from one file and then use it to build the local project.

If there is a File.Watcher, it can be kept around, though it’s not
clear if it is really worth keeping it in memory. Need to see how many
MB it tends to be / how much of a perf difference this makes.