Revamp File.Plan to load interfaces more efficiently
At this point in the pipeline we know that the dependencies form a DAG,
so we spawn each node as a process that blocks on its dependencies.

Each node will check if (1) all of its dependencies are fresh and (2)
if there is a cached interface for the source. If so, it is clean and
we filter it out. If not, we start loading all the dependencies we can
(which will rely on Haskell’s MIO
http://haskell.cs.yale.edu/wp-content/uploads/2013/08/hask035-voellmy.pd
f behind the scenes) for use in the upcoming compile phase.