Rewrite File.Compile to use new MVar approach
This is not fully working yet, but it outlines the core concepts.

Like in File.Plan we spawn all of the dirty nodes at once. They will
block until all of their dependencies are ready. At that point they
start compiling. This means all of the bookkeeping from the old
approach is gone, so we avoid a bunch of O(log n) lookups and updates
because we have no centralized manager at all. When something is done,
the relevant threads will wake up!

Remaining work:
  - Should compilation be in a thread pool?
  - Figure out reporting.
  - Figure out how to manage .elmo files