Add beginnings of concurrent package downloader
However big your thread pool is, that is how many downloads can be in
flight at the same time. Would be cool to start the per package
verification as soon as the relevant MVars were ready rather than
waiting for all of them. Suggests that requests should happen in
topological order.