Give each file all the available interfaces
The interfaces are kept in an MVar that is shared between all the
threads. Whenever there is an interface to add, it TAKES the value,
blocking all the other threads. This way we guarantee that no
interfaces are lost by two threads trying to change the value at the
same time.