Build dependencies after download / simplify lock file
Deps.Verify will now download and build stuff. The resulting interfaces
are consolidated into elm-stuff/0.18.1/summary.dat so that you only
need to read one file to get everything.

Whenever interfaces are written to disk, they are “privatized” and
minified by cutting out any information that is not needed outside of
the package itself.

This version blocks on reading all of summary.dat. May be worth
revisiting at some point.

The lock file is now just a (Map Name Version) now. That seems to be
all that is strictly necessary to verify that the cache is still valid.
Need to decide if this should be JSON or if it should be combined with
summary.dat or something else.