Deps.Verify can now skip builds if already done
When verifying a package build, you must build the particular
dependencies used in a project. We now record successful cases in
cached.dat so that in future situations, we can skip the build if we
have done it once already.

This change also collapses Deps.Interface into Deps.Verify so all the
related stuff is in one place.