Refactor Elm.Project.* and add Json.Encode
The Elm.Project.Root is now the entry point for getting project
information. Running Root.get will:

  - Find the nearest elm.json and move to that directory
  - Verify that all the necessary dependencies are ready
  - Create a “summary” that is used in later stages of the pipeline

This change also adds Elm.Project.Root.getWithReplFallback which looks
for the nearest elm.json file. If one is found, it moves there. If one
is not found, it moves to ~/.elm/repl/tmp/ and creates a dummy elm.json.

This change also moves decoders to Json.Decode and adds a Json.Encode
module that produces bytestring builders directly, resulting in pretty
output that should be quite fast.