Store source in graph / add crawlFromSource for REPL
This way each file is only read once. Things are kept in memory. This
gets filtered down in the Plan phase to only dirty files, and then
again to zero in the compile phase when it’s converted to AST.

The File.Crawl.crawlFromSource function will be helpful for the REPL
and for the /try page, allowing the source to be acquired by
alternative means.