Start revamping File.Generate to handle elm-build-plan.json
The old version relied on Data.Graph to do a topological sort. The new
one has a more direct implementation that I think is clearer. You do
not need to go figure out Data.Graph at least. It can also be modified
to allow streaming if that seems important.

There is also code to write all .elmo information into one file. This
code has very low memory overhead and computes a hash for the resulting
file. That means the resulting file name is a hash of its content.