Allow offline dependency solving if cannot get new packages
This will show a warning on the console in case it is a persistent
issue. If you make it past loading new packages, but then need to
download metadata, things will fail then.

This means you can start new projects offline as long as all the
dependencies have been downloaded before.