Update for new public API of elm-compiler
Highlights include:

1. No more Generate.App. Generating exports now happens in the
compiler, and then is hooked into JS by some code in core. This means
the exact behavior of that can change in a release of core.

2. No more Generate.Repl. More stuff is done in elm-compiler to
generate the footer that prints things out. This means I do not need to
leak any details about variable naming conventions outside of
elm-compiler.

3. The File.IO.Writer API no longer uses ReaderT. It is just a (Handle
-> IO ()) function! It may not actually be necessary though. Need to
check.

Otherwise it is mostly shuffling for API changes.